+++ rustyhack_lib/src/ecs/monster.rs
@@ -57,6 +57,7 @@ impl Default for Monster {
                 level: 1,
                 exp: 0,
                 exp_next: 1000,
+                in_combat: false,
             },
             inventory: Inventory::default(),
         }