+++ rustyhack_server/src/game.rs
@@ -34,9 +34,8 @@ pub(crate) fn run(sender: Sender<Packet>, receiver: Receiver<SocketEvent>) {
     let (default_spawn_counts, all_spawns_map) = spawns::initialise_all_spawn_definitions();
     let mut world = World::default();
     info!("Initialised ECS World");
-    let mut normal_player_update_schedule = systems::build_normal_player_update_schedule();
-    let mut player_combat_update_schedule = systems::build_player_combat_update_schedule();
-    let mut monster_update_schedule = systems::build_monster_update_schedule();
+    let mut player_update_schedule = systems::build_player_update_schedule();
+    let mut server_tick_update_schedule = systems::build_server_tick_update_schedule();
     let mut map_state_update_schedule = systems::build_map_state_update_schedule();
     let mut health_regen_schedule = systems::build_health_regen_schedule();
     let mut send_network_messages_schedule = systems::send_network_messages_schedule();
@@ -79,7 +78,7 @@ pub(crate) fn run(sender: Sender<Packet>, receiver: Receiver<SocketEvent>) {
         if player_updates::process_player_messages(&mut world, &channel_receiver, &local_sender) {
             debug!("Executing player update schedule...");
             map_state_update_schedule.execute(&mut world, &mut resources);
-            normal_player_update_schedule.execute(&mut world, &mut resources);
+            player_update_schedule.execute(&mut world, &mut resources);
             debug!("Player update schedule executed successfully.");
 
             //send game tick updates to players
@@ -89,11 +88,10 @@ pub(crate) fn run(sender: Sender<Packet>, receiver: Receiver<SocketEvent>) {
         //all other updates that depend on the server game tick
         if server_game_tick_time.elapsed() > consts::SERVER_GAME_TICK {
             server_game_tick_count += 1;
-            debug!("Executing monster and player combat update schedule...");
+            debug!("Executing server tick schedule...");
             map_state_update_schedule.execute(&mut world, &mut resources);
-            player_combat_update_schedule.execute(&mut world, &mut resources);
-            monster_update_schedule.execute(&mut world, &mut resources);
-            debug!("Monster and player combat update schedule executed successfully.");
+            server_tick_update_schedule.execute(&mut world, &mut resources);
+            debug!("Server tick update schedule executed successfully.");
 
             //things that happen every 2 ticks rather than every tick
             if server_game_tick_count == 2 {