+++ rustyhack_server/src/game/systems/combat_systems.rs
@@ -1,4 +1,4 @@
-use crate::consts::MONSTER_EXP_MULTIPLICATION_FACTOR;
+use crate::consts::{GOLD_LOSS_ON_PVP_DEATH_PERCENTAGE, MONSTER_EXP_MULTIPLICATION_FACTOR};
 use crate::game::combat::{Attacker, CombatAttackerStats, CombatParties, Defender};
 use crate::game::map_state::{AllMapStates, MapState};
 use crate::game::{combat, map_state};
@@ -24,7 +24,7 @@ pub(crate) fn check_for_combat(
         Option<&MonsterDetails>,
         Option<&PlayerDetails>,
         &DisplayDetails,
-        &Stats,
+        &mut Stats,
         &Inventory,
     )>,
     #[resource] all_map_states: &mut AllMapStates,
@@ -41,8 +41,9 @@ pub(crate) fn check_for_combat(
     ) in query.iter_mut(world)
     {
         debug!("Checking for possible combat after velocity updates.");
-        if position.velocity_x == 0 && position.velocity_y == 0 {
-            //no velocity, no updates
+        if (position.velocity_x == 0 && position.velocity_y == 0) || stats.current_hp <= 0.0 {
+            //either not moving, or already dead, no combat
+            stats.in_combat = false;
             continue;
         }
         let current_map_state = get_current_map_states(all_map_states, &position.current_map);
@@ -58,22 +59,26 @@ pub(crate) fn check_for_combat(
         let attacker = get_attacker(player_details_option, monster_details_option);
 
         if entity_collision_status.0 {
+            //combat detected
             if !attacker.is_player && !entity_collision_status.1.is_player {
                 debug!("Preventing possible monster combat.");
                 //monsters shouldn't attack other monsters, stop combat and movement
                 position.velocity_x = 0;
                 position.velocity_y = 0;
+                stats.in_combat = false;
             } else {
                 debug!(
                     "Combat detected, attacker is: {:?}, defender is: {:?}",
                     &attacker, &entity_collision_status.1
                 );
-                combat_parties.insert(entity_collision_status.1, attacker.clone());
+                combat_parties.insert(attacker.clone(), entity_collision_status.1);
                 combat_attacker_stats.insert(attacker.id, (*stats, inventory.clone()));
                 position.velocity_x = 0;
                 position.velocity_y = 0;
+                stats.in_combat = true;
             }
         } else if attacker.is_player {
+            //no combat detected - update map state for movement
             let player = Player {
                 player_details: player_details_option.unwrap().clone(),
                 display_details: *display_details,
@@ -93,7 +98,13 @@ pub(crate) fn check_for_combat(
                 position.pos_x,
                 position.pos_y,
             );
+            //player has moved, therefore no longer attacking
+            if combat_parties.contains_key(&attacker) {
+                combat_parties.remove(&attacker);
+            }
+            stats.in_combat = false;
         } else if !attacker.is_player {
+            //no combat detected - update map state for movement
             let monster = Monster {
                 monster_details: monster_details_option.unwrap().clone(),
                 display_details: *display_details,
@@ -113,6 +124,11 @@ pub(crate) fn check_for_combat(
                 position.pos_x,
                 position.pos_y,
             );
+            //monster has moved, therefore no longer attacking
+            if combat_parties.contains_key(&attacker) {
+                combat_parties.remove(&attacker);
+            }
+            stats.in_combat = false;
         }
     }
 }
@@ -178,6 +194,7 @@ pub(crate) fn resolve_combat(
             continue;
         }
         let mut defender_is_monster = false;
+        let mut monster_target = None;
         let mut defender: Defender = Defender::default();
         if let Some(player_details) = player_details_option {
             //player is the defender
@@ -197,30 +214,38 @@ pub(crate) fn resolve_combat(
                 currently_online: true,
                 is_player: false,
             };
+            monster_target = monster_details.current_target;
             defender_is_monster = true;
         }
-        if combat_parties.contains_key(&defender) {
-            let attacker = combat_parties.get(&defender).unwrap();
-            let (mut attacker_stats, mut attacker_inventory) =
-                combat_attacker_stats.get(&attacker.id).unwrap().clone();
-            if attacker_stats.current_hp <= 0.0 {
-                // Skip combat if attacker is already dead.
-                // This is possible if multiple player updates are processed
-                // before the server tick for monsters.
-                continue;
-            }
-            let damage = combat::resolve_combat(
-                &attacker_stats,
-                &attacker_inventory,
-                defender_stats,
-                defender_inventory,
-            )
-            .round();
-            apply_damage(defender_stats, damage);
-            let mut exp_gain = 0;
-            let mut gold_gain = 0;
-            if defender_is_monster {
-                (exp_gain, gold_gain) = check_and_apply_gains(
+        for (combat_parties_attacker, combat_parties_defender) in combat_parties.iter() {
+            if combat_parties_defender == &defender {
+                //combat detected
+                defender_stats.in_combat = true;
+                let attacker = combat_parties_attacker;
+                let (mut attacker_stats, mut attacker_inventory) =
+                    combat_attacker_stats.get(&attacker.id).unwrap().clone();
+                if attacker_stats.current_hp <= 0.0 {
+                    // Skip combat if attacker is already dead.
+                    // This is possible if combat is happening simultaneously.
+                    continue;
+                }
+                let damage = combat::resolve_combat(
+                    &attacker_stats,
+                    &attacker_inventory,
+                    defender_stats,
+                    defender_inventory,
+                )
+                .round();
+                apply_damage(defender_stats, damage);
+                if combat_attacker_stats.contains_key(&defender.id) {
+                    combat_attacker_stats
+                        .get_mut(&defender.id)
+                        .unwrap()
+                        .0
+                        .current_hp = defender_stats.current_hp;
+                }
+                let (exp_gain, gold_gain) = check_and_apply_gains(
+                    defender_is_monster,
                     combat_attacker_stats,
                     &attacker.id,
                     &mut attacker_stats,
@@ -228,45 +253,66 @@ pub(crate) fn resolve_combat(
                     defender_stats,
                     defender_inventory,
                 );
-                //set current monster target as attacker
-                monster_details_option.unwrap().current_target = Some(attacker.id);
+                if defender_is_monster {
+                    //set current monster target as attacker
+                    monster_target = Some(attacker.id);
+                }
+                combat::send_combat_updates_to_players(
+                    &defender,
+                    attacker,
+                    damage,
+                    defender_stats.current_hp,
+                    exp_gain,
+                    gold_gain,
+                    sender,
+                );
+
+                if let Some(_player_details) = player_details_option {
+                    //only set flag for players
+                    defender_stats.update_available = true;
+                }
             }
-            combat::send_combat_updates_to_players(
-                &defender,
-                attacker,
-                damage,
-                defender_stats.current_hp,
-                exp_gain,
-                gold_gain,
-                sender,
-            );
-            if let Some(_player_details) = player_details_option {
-                //only set flag for players
-                defender_stats.update_available = true;
-            }
+        }
+        if defender_is_monster {
+            //set current_target to most recent attacker
+            monster_details_option.unwrap().current_target = monster_target;
         }
     }
     combat_parties.clear();
 }
 
 fn check_and_apply_gains(
+    defender_is_monster: bool,
     combat_attacker_stats: &mut CombatAttackerStats,
     attacker_id: &Uuid,
     attacker_stats: &mut Stats,
     attacker_inventory: &mut Inventory,
     defender_stats: &Stats,
-    defender_inventory: &Inventory,
+    defender_inventory: &mut Inventory,
 ) -> (u32, u32) {
     if defender_stats.current_hp <= 0.0 {
         //calculate xp to be gained
-        let exp_gain = defender_stats.level * MONSTER_EXP_MULTIPLICATION_FACTOR;
-        attacker_stats.exp += exp_gain;
-        attacker_stats.update_available = true;
+        let mut exp_gain = 0;
+        if defender_is_monster {
+            exp_gain = defender_stats.level * MONSTER_EXP_MULTIPLICATION_FACTOR;
+            attacker_stats.exp += exp_gain;
+        }
 
         //calculate gold to be gained
-        let gold_gain = defender_inventory.gold;
+        let mut gold_gain = 0;
+        if defender_is_monster {
+            gold_gain = defender_inventory.gold;
+        } else if defender_inventory.gold > 100 {
+            gold_gain = (defender_inventory.gold * GOLD_LOSS_ON_PVP_DEATH_PERCENTAGE) / 100;
+            defender_inventory.gold -= gold_gain;
+            defender_inventory.update_available = true;
+        }
         attacker_inventory.gold += gold_gain;
+        if exp_gain > 0 || gold_gain > 0 {
+            attacker_stats.update_available = true;
+        }
         combat_attacker_stats.insert(*attacker_id, (*attacker_stats, attacker_inventory.clone()));
+
         return (exp_gain, gold_gain);
     }
     (0, 0)